requirements.txt
aiomysql==0.2.0
annotated-types==0.7.0
bcrypt==4.3.0
bson==0.5.10
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==44.0.3
dependency-injector==4.46.0
dnspython==2.7.0
email_validator==2.2.0
filelock==3.18.0
idna==3.10
psutil==7.0.0
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
requests-file==2.1.0
six==1.17.0
slugify==0.0.1
tldextract==5.3.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
